+++ backend/src/HomeService.Application/Features/Reviews/CreateReviewCommand.cs
@@ -153,7 +153,7 @@ public class CreateReviewCommandHandler : IRequestHandler<CreateReviewCommand, R
 
             if (provider != null)
             {
-                provider.AverageRating = Math.Round(averageRating, 2);
+                provider.AverageRating = (decimal)Math.Round(averageRating, 2);
                 provider.TotalReviews = reviews.Count;
                 _unitOfWork.Repository<ServiceProvider>().Update(provider);
                 await _unitOfWork.SaveChangesAsync(cancellationToken);
