+++ backend/src/HomeService.Application/Features/Reviews/DeleteReviewCommand.cs
@@ -75,7 +75,7 @@ public class DeleteReviewCommandHandler : IRequestHandler<DeleteReviewCommand, R
         {
             if (reviews.Any())
             {
-                provider.AverageRating = Math.Round(reviews.Average(r => r.Rating), 2);
+                provider.AverageRating = (decimal)Math.Round(reviews.Average(r => r.Rating), 2);
                 provider.TotalReviews = reviews.Count;
             }
             else
